issues in saga resolved
