+++ frontend/src/utils/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from "@reduxjs/toolkit";
-import studentReducer from "../components/DataGrid/studentSlice";
-import authReducer from "../pages/SignInPage/authSlice";
+import studentReducer from "../pages/DataGridPage/slice";
+import authReducer from "../pages/SignInPage/slice";
 
 const rootReducer = combineReducers({
   student: studentReducer,
