+++ frontend/src/utils/rootSaga.ts
@@ -1,7 +1,7 @@
 import { all } from "redux-saga/effects";
-import studentSaga from "../components/DataGrid/studentSaga";
+import studentSaga from "../pages/DataGridPage/saga";
 import notificationSaga from "./notificationSaga";
-import authSaga from "../pages/SignInPage/authSaga";
+import authSaga from "../pages/SignInPage/saga";
 
 export default function* rootSaga() {
   yield all([studentSaga(), notificationSaga(), authSaga()]);
